Refactored Units Tests
