+++ test/Lokad.ILPack.Tests/RewriteTest.Async.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Xunit;
+
+
+namespace Lokad.ILPack.Tests
+{
+    partial class RewriteTest
+    {
+        /*
+        [Fact]
+        public async void AsyncMethod()
+        {
+            Assert.Equal(60, await Invoke(
+                $"int r = await x.AsyncMethod(30, 30);",
+                "r"));
+        }
+        */
+    }
+}
