+++ test/Lokad.ILPack.Tests/RewriteTest.Enums.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Xunit;
+
+
+namespace Lokad.ILPack.Tests
+{
+    partial class RewriteTest
+    {
+        [Fact]
+        public async void BasicEnumTest()
+        {
+            Assert.Equal(150, await Invoke(
+                $"int r = (int)MyEnum.Pears;",
+                "r"));
+        }
+
+    }
+}
