+++ test/Lokad.ILPack.Tests/RewriteTest.Properties.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Xunit;
+
+namespace Lokad.ILPack.Tests
+{
+    partial class RewriteTest
+    {
+        [Fact]
+        public async void ReadOnlyProperty()
+        {
+            Assert.Equal(23, await Invoke(
+                "",
+                "x.ReadOnlyProperty"));
+        }
+
+        [Fact]
+        public async void WriteOnlyProperty()
+        {
+            Assert.Equal(true, await Invoke(
+                "x.WriteOnlyProperty = 99;",
+                "true"));
+        }
+
+        [Fact]
+        public async void ReadWriteOnlyProperty()
+        {
+            Assert.Equal(101, await Invoke(
+                "x.ReadWriteProperty = 101;",
+                "x.ReadWriteProperty"));
+        }
+    }
+}
