+++ test/TestSubject/MyClass.Indexer.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Threading.Tasks;
+
+// This project defines a set of types that will be rewritten by the test cases to a new
+// dll named "ClonedTestSubject".  The test cases then compare the final type information of both
+// assemblies to confirm everything was re-written correction
+
+// SEE RewriteTest in Lokad.ILPack.Tests
+
+
+namespace TestSubject
+{
+    public partial class MyClass
+    {
+        int[] _array = new int[10];
+        public int this[int index]
+        {
+            get => _array[index];
+            set => _array[index] = value;
+        }
+    }
+}
