+++ test/TestSubject/MyClass.Structs.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Threading.Tasks;
+
+// This project defines a set of types that will be rewritten by the test cases to a new
+// dll named "ClonedTestSubject".  The test cases then compare the final type information of both
+// assemblies to confirm everything was re-written correction
+
+// SEE RewriteTest in Lokad.ILPack.Tests
+
+
+namespace TestSubject
+{
+    public partial class MyClass
+    {
+        public MyStruct GetMyStruct()
+        {
+            return new MyStruct()
+            {
+                x = 20,
+                y = 30,
+            };
+        }
+    }
+}
